moving fighting folders to main page
