+++ README.md
@@ -1,6 +1,15 @@
-# Project-Fighting
+## Express Sequelize Handlebars Boilerplate
 
+### Setup
 
-## Hello World 
+1. clone repo
+2. npm install
+3. create db named blogger in mySQL workbench or SequelPro
+4. update the config file with your sql credentials
+5. node server.js
 
+This project renders views in two different ways so you can choose how you want to code!
+
+- The blog template uses handle bars to display data on the page server-side
+- The authors and cms templates use jQuery ajax to retrieve data on the client, and then builds the view with jQuery dom manipulation
