--- controllers/workoutController.js
@@ -1,22 +1,0 @@
-var express = require("express");
-
-var router = express.Router();
-
-// Import the model (cat.js) to use its database functions.
-var workout = require("../models/workout.js");
-
-// Create all our routes and set up logic within those routes where required.
-router.get("/", function (req, res) {
-    workout.all(function (data) {
-        var hbsObject = {
-            cats: data
-        };
-        console.log(hbsObject);
-        res.render("index", hbsObject);
-    });
-});
-
-
-
-// Export routes for server.js to use.
-module.exports = router;
