--- fighting/.eslintignore
@@ -1,2 +1,0 @@
-models/index.js
-node_modules
