--- fighting/.travis.yml
@@ -1,14 +1,0 @@
-language: node_js
-node_js: "12"
-branches:
-  only:
-    - master
-    - seulkiBranch
-    - k-branch
-cache:
-  directories:
-    - node_modules
-services:
-  - mysql
-before_install:
-  - mysql -e 'CREATE DATABASE IF NOT EXISTS testdb;'
