--- fighting/db/schema.sql
@@ -1,4 +1,0 @@
--- Drops the workout if it exists currently --
-DROP DATABASE IF EXISTS workoutdb;
--- Creates the "workout" database --
-CREATE DATABASE workoutdb;
