--- fighting/public/assets/css/style.css
@@ -1,13 +1,0 @@
-  
-.list-group-item {
-  line-height: 2.5;
-}
-
-.navbar-title {
-  color: #4a4a4a;
-}
-
-.navbar-title:hover {
-  text-decoration: none;
-  color: #4a4a4a;
-}
