--- fighting/views/index.handlebars
@@ -1,28 +1,0 @@
-<div id="home">
-
-    <section class="hero is-dark is-medium">
-        <div class="hero-body has-text-centered">
-            <div class="container">
-                <h1 class="title is-1 is-spaced">
-                    Welcome to Workout Tracker!
-                </h1>
-                <h2 class="subtitle is-4">
-                    Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the
-                    industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type
-                    and scrambled it to make a type specimen book.
-                </h2>
-            </div>
-        </div>
-        <div class="hero-footer has-text-centered" style="padding-bottom: 2rem;">
-            <div class="container buttons is-centered">
-                <button class="button is-light is-large" style="margin-left: 5 rem; margin-right: 5rem;"><a
-                        href="./">View Your Workout</a></button>
-                <button class="button is-light is-large" style="margin-left: 5 rem; margin-right: 5rem;"><a
-                        href="./">Add Workout</a></button>
-                <button class="button is-light is-large" style="margin-left: 5 rem; margin-right: 5rem;"><a
-                        href="./">Discover</a></button>
-            </div>
-        </div>
-    </section>
-
-</div>
