+++ models/workout.js
@@ -1,24 +1,17 @@
-// Import the ORM to create functions that will interact with the database.
-var orm = require("../config/orm.js");
+module.exports = function (sequelize, DataTypes) {
 
-var workout = {
-    all: function (cb) {
-        orm.all("workouts", function (res) {
-            cb(res);
+    var Workout = sequelize.define("Workout", {
+        // Giving the Workout model a name of type STRING
+        name: DataTypes.STRING
+    });
+
+    Workout.associate = function (models) {
+        // Associating Workout with Exercise
+        // When a Workout is deleted, also delete any associated Exercise
+        Workout.hasMany(models.Exercise, {
+            onDelete: "cascade"
         });
-    },
-    // The variables cols and vals are arrays.
-    create: function (cols, vals, cb) {
-        orm.create("workouts", cols, vals, function (res) {
-            cb(res);
-        });
-    },
-    update: function (objColVals, condition, cb) {
-        orm.update("workouts", objColVals, condition, function (res) {
-            cb(res);
-        });
-    }
+    };
+
+    return Workout;
 };
-
-// Export the database functions for the controller (catsController.js).
-module.exports = workout;
