+++ package.json
@@ -1,28 +1,21 @@
 {
-  "name": "project-fighting",
+  "name": "post-author-association",
   "version": "1.0.0",
-  "description": "Project 2",
+  "description": "",
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "start": "node server.js",
+    "watch": "nodemon server.js"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/seulkilee0428/Project-Fighting.git"
-  },
-  "author": "kurtiam, seulkile0428",
+  "keywords": [],
+  "author": "",
   "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/seulkilee0428/Project-Fighting/issues"
-  },
-  "homepage": "https://github.com/seulkilee0428/Project-Fighting#readme",
   "dependencies": {
-    "eslint": "^7.4.0",
-    "express": "^4.17.1",
-    "express-handlebars": "^5.0.0",
-    "mysql": "^2.18.1",
-    "mysql2": "^2.1.0",
-    "sequelize": "^6.3.0"
+    "body-parser": "^1.15.2",
+    "express": "^4.14.0",
+    "express-handlebars": "^3.0.0",
+    "mysql2": "^1.3.6",
+    "sequelize": "^4.3.1"
   }
 }
