+++ server.js
@@ -1,15 +1,29 @@
+// *****************************************************************************
+// Server.js - This file is the initial starting point for the Node/Express server.
+//
+// ******************************************************************************
+// *** Dependencies
+// =============================================================
 var express = require("express");
+var bodyParser = require("body-parser");
 
+// Sets up the Express App
+// =============================================================
+var app = express();
 var PORT = process.env.PORT || 8080;
 
-var app = express();
+// Requiring our models for syncing
+var db = require("./models");
 
-// Serve static content for the app from the "public" directory in the application directory.
+// Sets up the Express app to handle data parsing
+
+// parse application/x-www-form-urlencoded
+app.use(bodyParser.urlencoded({ extended: false }));
+// parse application/json
+app.use(bodyParser.json());
+
+// Static directory
 app.use(express.static("public"));
-
-// Parse application body as JSON
-app.use(express.urlencoded({ extended: true }));
-app.use(express.json());
 
 // Set Handlebars.
 var exphbs = require("express-handlebars");
@@ -17,13 +31,21 @@ var exphbs = require("express-handlebars");
 app.engine("handlebars", exphbs({ defaultLayout: "main" }));
 app.set("view engine", "handlebars");
 
-// Import routes and give the server access to them.
-var routes = require("./controllers/workoutController.js");
+// Routes
+// =============================================================
+const viewController = require("./controllers/view-controller.js");
+const workoutController = require("./controllers/workout-controller.js");
+const exerciseController = require("./controllers/exercise-controller.js");
 
-app.use(routes);
 
-// Start our server so that it can begin listening to client requests.
-app.listen(PORT, function () {
-  // Log (server-side) when our server has started
-  console.log("Server listening on: http://localhost:" + PORT);
+app.use(viewController);
+app.use(workoutController);
+app.use(exerciseController);
+
+// Syncing our sequelize models and then starting our Express app
+// =============================================================
+db.sequelize.sync().then(function () {
+  app.listen(PORT, function () {
+    console.log("App listening on PORT " + PORT);
+  });
 });
